+++ src/Giraffe/Routing.fs
@@ -13,33 +13,36 @@ open Giraffe.FormatExpressions
 // Private sub route helper functions
 // ---------------------------
 
-[<Literal>]
-let private RouteKey = "giraffe_route"
+[<RequireQualifiedAccess>]
+module SubRouting =
 
-let private getSavedSubPath (ctx : HttpContext) =
-    if ctx.Items.ContainsKey RouteKey
-    then ctx.Items.Item RouteKey |> string |> strOption
-    else None
+    [<Literal>]
+    let private RouteKey = "giraffe_route"
 
-let private getPath (ctx : HttpContext) =
-    match getSavedSubPath ctx with
-    | Some p when ctx.Request.Path.Value.Contains p -> ctx.Request.Path.Value.[p.Length..]
-    | _   -> ctx.Request.Path.Value
+    let getSavedPartialPath (ctx : HttpContext) =
+        if ctx.Items.ContainsKey RouteKey
+        then ctx.Items.Item RouteKey |> string |> strOption
+        else None
 
-let private handlerWithRootedPath (path : string) (handler : HttpHandler) : HttpHandler =
-    fun (next : HttpFunc) (ctx : HttpContext) ->
-        task {
-            let savedSubPath = getSavedSubPath ctx
-            ctx.Items.Item RouteKey <- ((savedSubPath |> Option.defaultValue "") + path)
-            let! result = handler next ctx
-            match result with
-            | Some _ -> ()
-            | None ->
-                match savedSubPath with
-                | Some savedSubPath -> ctx.Items.Item   RouteKey <- savedSubPath
-                | None              -> ctx.Items.Remove RouteKey |> ignore
-            return result
-        }
+    let getNextPartOfPath (ctx : HttpContext) =
+        match getSavedPartialPath ctx with
+        | Some p when ctx.Request.Path.Value.Contains p -> ctx.Request.Path.Value.[p.Length..]
+        | _   -> ctx.Request.Path.Value
+
+    let routeWithPartialPath (path : string) (handler : HttpHandler) : HttpHandler =
+        fun (next : HttpFunc) (ctx : HttpContext) ->
+            task {
+                let savedPartialPath = getSavedPartialPath ctx
+                ctx.Items.Item RouteKey <- ((savedPartialPath |> Option.defaultValue "") + path)
+                let! result = handler next ctx
+                match result with
+                | Some _ -> ()
+                | None ->
+                    match savedPartialPath with
+                    | Some subPath -> ctx.Items.Item   RouteKey <- subPath
+                    | None         -> ctx.Items.Remove RouteKey |> ignore
+                return result
+            }
 
 // ---------------------------
 // Public routing HttpHandler functions
@@ -84,7 +87,7 @@ let routePorts (fns : (int * HttpHandler) list) : HttpHandler =
 ///
 let route (path : string) : HttpHandler =
     fun (next : HttpFunc) (ctx : HttpContext) ->
-        if (getPath ctx).Equals path
+        if (SubRouting.getNextPartOfPath ctx).Equals path
         then next ctx
         else abort
 
@@ -102,7 +105,7 @@ let route (path : string) : HttpHandler =
 ///
 let routeCi (path : string) : HttpHandler =
     fun (next : HttpFunc) (ctx : HttpContext) ->
-        if String.Equals(getPath ctx, path, StringComparison.CurrentCultureIgnoreCase)
+        if String.Equals(SubRouting.getNextPartOfPath ctx, path, StringComparison.CurrentCultureIgnoreCase)
         then next ctx
         else abort
 
@@ -122,7 +125,7 @@ let routex (path : string) : HttpHandler =
     fun (next : HttpFunc) (ctx : HttpContext) ->
         let pattern = sprintf "^%s$" path
         let regex   = Regex(pattern, RegexOptions.Compiled)
-        let result  = regex.Match (getPath ctx)
+        let result  = regex.Match (SubRouting.getNextPartOfPath ctx)
         match result.Success with
         | true -> next ctx
         | false -> abort
@@ -143,7 +146,7 @@ let routeCix (path : string) : HttpHandler =
     fun (next : HttpFunc) (ctx : HttpContext) ->
         let pattern = sprintf "^%s$" path
         let regex   = Regex(pattern, RegexOptions.IgnoreCase ||| RegexOptions.Compiled)
-        let result  = regex.Match (getPath ctx)
+        let result  = regex.Match (SubRouting.getNextPartOfPath ctx)
         match result.Success with
         | true -> next ctx
         | false -> abort
@@ -176,7 +179,7 @@ let routeCix (path : string) : HttpHandler =
 let routef (path : PrintfFormat<_,_,_,_, 'T>) (routeHandler : 'T -> HttpHandler) : HttpHandler =
     validateFormat path
     fun (next : HttpFunc) (ctx : HttpContext) ->
-        tryMatchInput path (getPath ctx) false
+        tryMatchInput path (SubRouting.getNextPartOfPath ctx) false
         |> function
             | None      -> abort
             | Some args -> routeHandler args next ctx
@@ -209,7 +212,7 @@ let routef (path : PrintfFormat<_,_,_,_, 'T>) (routeHandler : 'T -> HttpHandler)
 let routeCif (path : PrintfFormat<_,_,_,_, 'T>) (routeHandler : 'T -> HttpHandler) : HttpHandler =
     validateFormat path
     fun (next : HttpFunc) (ctx : HttpContext) ->
-        tryMatchInput path (getPath ctx) true
+        tryMatchInput path (SubRouting.getNextPartOfPath ctx) true
         |> function
             | None      -> abort
             | Some args -> routeHandler args next ctx
@@ -233,7 +236,7 @@ let routeBind<'T> (route : string) (routeHandler : 'T -> HttpHandler) : HttpHand
     fun (next : HttpFunc) (ctx : HttpContext) ->
         let pattern = route.Replace("{", "(?<").Replace("}", ">[^/\n]+)") |> sprintf "^%s$"
         let regex   = Regex(pattern, RegexOptions.IgnoreCase)
-        let result  = regex.Match (getPath ctx)
+        let result  = regex.Match (SubRouting.getNextPartOfPath ctx)
         match result.Success with
         | true ->
             let groups = result.Groups
@@ -262,7 +265,7 @@ let routeBind<'T> (route : string) (routeHandler : 'T -> HttpHandler) : HttpHand
 ///
 let routeStartsWith (subPath : string) : HttpHandler =
     fun (next : HttpFunc) (ctx : HttpContext) ->
-        if (getPath ctx).StartsWith subPath
+        if (SubRouting.getNextPartOfPath ctx).StartsWith subPath
         then next ctx
         else abort
 
@@ -280,7 +283,7 @@ let routeStartsWith (subPath : string) : HttpHandler =
 ///
 let routeStartsWithCi (subPath : string) : HttpHandler =
     fun (next : HttpFunc) (ctx : HttpContext) ->
-        if (getPath ctx).StartsWith(subPath, StringComparison.CurrentCultureIgnoreCase)
+        if (SubRouting.getNextPartOfPath ctx).StartsWith(subPath, StringComparison.CurrentCultureIgnoreCase)
         then next ctx
         else abort
 
@@ -300,7 +303,7 @@ let routeStartsWithCi (subPath : string) : HttpHandler =
 ///
 let subRoute (path : string) (handler : HttpHandler) : HttpHandler =
     routeStartsWith path >=>
-    handlerWithRootedPath path handler
+    SubRouting.routeWithPartialPath path handler
 
 /// **Description**
 ///
@@ -318,7 +321,7 @@ let subRoute (path : string) (handler : HttpHandler) : HttpHandler =
 ///
 let subRouteCi (path : string) (handler : HttpHandler) : HttpHandler =
     routeStartsWithCi path >=>
-    handlerWithRootedPath path handler
+    SubRouting.routeWithPartialPath path handler
 
 /// **Description**
 ///
@@ -351,7 +354,7 @@ let subRoutef (path : PrintfFormat<_,_,_,_, 'T>) (routeHandler : 'T -> HttpHandl
         validateFormat path
         fun (next : HttpFunc) (ctx : HttpContext) ->
             let paramCount   = (path.Value.Split '/').Length
-            let subPathParts = (getPath ctx).Split '/'
+            let subPathParts = (SubRouting.getNextPartOfPath ctx).Split '/'
             if paramCount > subPathParts.Length then abort
             else
                 let subPath =
@@ -364,4 +367,4 @@ let subRoutef (path : PrintfFormat<_,_,_,_, 'T>) (routeHandler : 'T -> HttpHandl
                 tryMatchInput path subPath false
                 |> function
                     | None      -> abort
-                    | Some args -> handlerWithRootedPath subPath (routeHandler args) next ctx
+                    | Some args -> SubRouting.routeWithPartialPath subPath (routeHandler args) next ctx